pruned cortex things out of snowflake data
