--- actions/snowflake-cortex/devdata/input_cortex_analyst_message.json
@@ -1,27 +1,0 @@
-{
-    "inputs": [
-        {
-            "inputName": "input-1",
-            "inputValue": {
-                "message": "Show me the top 5 oil producing wells by production volume over time.",
-                "semantic_model_file": "@PRODUCTION_RESULTS.PUBLIC.STAGE1/oil_gas.yml"
-            }
-        }
-    ],
-    "metadata": {
-        "actionName": "cortex_analyst_message",
-        "actionRelativePath": "cortex_actions.py",
-        "schemaDescription": [
-            "message: string: The message to send to the Cortex Analyst."
-        ],
-        "managedParamsSchemaDescription": {
-            "semantic_model_file": {
-                "type": "Secret",
-                "description": "The semantic model file to use."
-            }
-        },
-        "inputFileVersion": "v3",
-        "kind": "action",
-        "actionSignature": "action/args: 'semantic_model_file: Secret, message: str'"
-    }
-}
