--- actions/snowflake-cortex/devdata/input_cortex_search.json
@@ -1,51 +1,0 @@
-{
-    "inputs": [
-        {
-            "inputName": "input-1",
-            "inputValue": {
-                "query": "Can you analyze the different types of cancellation policies by room type?",
-                "columns": [
-                    "room_type",
-                    "cancellation_policy"
-                ],
-                "limit": 10,
-                "warehouse": "COMPUTE_WH",
-                "database": "CORTEX_SEARCH_TUTORIAL_DB",
-                "schema": "PUBLIC",
-                "service": "AIRBNB_SVC"
-            }
-        }
-    ],
-    "metadata": {
-        "actionName": "cortex_search",
-        "actionRelativePath": "cortex_actions.py",
-        "schemaDescription": [
-            "query: string: The query to execute",
-            "columns: array: The columns to return",
-            "columns.0: string: The columns to return",
-            "filter: string: The filter to apply, optional, defaults to None",
-            "limit: integer: The limit to apply, optional, defaults to 5"
-        ],
-        "managedParamsSchemaDescription": {
-            "warehouse": {
-                "type": "Secret",
-                "description": "Your Snowflake virtual warehouse to use for queries"
-            },
-            "database": {
-                "type": "Secret",
-                "description": "Your Snowflake database to use for queries"
-            },
-            "schema": {
-                "type": "Secret",
-                "description": "Your Snowflake schema to use for queries"
-            },
-            "service": {
-                "type": "Secret",
-                "description": "The name of the Cortex Search service to use"
-            }
-        },
-        "inputFileVersion": "v3",
-        "kind": "action",
-        "actionSignature": "action/args: 'query: str, warehouse: Secret, database: Secret, schema: Secret, service: Secret, columns: list, filter: dict | None=None, limit: int=5'"
-    }
-}
