--- actions/snowflake-cortex/devdata/input_snowflake_get_schemas.json
@@ -1,23 +1,0 @@
-{
-    "inputs": [
-        {
-            "inputName": "input-1",
-            "inputValue": {
-                "warehouse": "PLATFORM_TEAM_DEFAULT",
-                "database": "RESOURCES_DEVELOPMENT"
-            }
-        }
-    ],
-    "metadata": {
-        "actionName": "snowflake_get_schemas",
-        "actionRelativePath": "snowflake_actions.py",
-        "schemaDescription": [
-            "warehouse: string: The warehouse to use.",
-            "database: string: The database to get the schemas from."
-        ],
-        "managedParamsSchemaDescription": {},
-        "inputFileVersion": "v3",
-        "kind": "action",
-        "actionSignature": "action/args: 'warehouse: str, database: str'"
-    }
-}
