+++ actions/snowflake-data/README.md
@@ -1,0 +1,17 @@
+# Snowflake Databases and Tools
+
+This action package provides seamless integration with Snowflake databases and other tools.
+
+## Available Actions
+
+### Snowflake Cortex
+- Cortex Agent Chat
+- Cortex LLM Complete
+
+### Snowflake Databases
+- Execute Query
+- Get Databases
+- Get Schemas
+- Get Tables
+- Get Columns
+- Get Warehouses
