+++ actions/snowflake-data/devdata/input_snowflake_execute_query.json
@@ -1,0 +1,39 @@
+{
+    "inputs": [
+        {
+            "inputName": "input-1",
+            "inputValue": {
+                "query": "SELECT * FROM daily_revenue LIMIT 10",
+                "warehouse": "COMPUTE_WH",
+                "database": "CORTEX_ANALYST_DEMO",
+                "schema": "REVENUE_TIMESERIES",
+                "numeric_args": null
+            }
+        }
+    ],
+    "metadata": {
+        "actionName": "snowflake_execute_query",
+        "actionRelativePath": "snowflake_actions.py",
+        "schemaDescription": [
+            "query: string: The query to execute.",
+            "numeric_args: string: A list of numeric arguments to pass to the query."
+        ],
+        "managedParamsSchemaDescription": {
+            "warehouse": {
+                "type": "Secret",
+                "description": "Your Snowflake virtual warehouse to use for queries."
+            },
+            "database": {
+                "type": "Secret",
+                "description": "Your Snowflake database to use for queries."
+            },
+            "schema": {
+                "type": "Secret",
+                "description": "Your Snowflake schema to use for queries."
+            }
+        },
+        "inputFileVersion": "v3",
+        "kind": "action",
+        "actionSignature": "action/args: 'query: str, warehouse: Secret, database: Secret, schema: Secret, numeric_args: list=None'"
+    }
+}
