+++ actions/snowflake-data/devdata/input_snowflake_get_schemas.json
@@ -3,19 +3,21 @@
         {
             "inputName": "input-1",
             "inputValue": {
-                "warehouse": "CORTEX_SEARCH_TUTORIAL_WH"
+                "warehouse": "COMPUTE_WH",
+                "database": "CORTEX_ANALYST_DEMO"
             }
         }
     ],
     "metadata": {
-        "actionName": "snowflake_get_databases",
+        "actionName": "snowflake_get_schemas",
         "actionRelativePath": "snowflake_actions.py",
         "schemaDescription": [
-            "warehouse: string: The warehouse to use."
+            "warehouse: string: The warehouse to use.",
+            "database: string: The database to get the schemas from."
         ],
         "managedParamsSchemaDescription": {},
         "inputFileVersion": "v3",
         "kind": "action",
-        "actionSignature": "action/args: 'warehouse: str'"
+        "actionSignature": "action/args: 'warehouse: str, database: str'"
     }
 }
