+++ actions/snowflake-data/devdata/input_snowflake_get_tables.json
@@ -3,9 +3,9 @@
         {
             "inputName": "input-1",
             "inputValue": {
-                "warehouse": "PLATFORM_TEAM_DEFAULT",
-                "database": "RESOURCES_DEVELOPMENT",
-                "schema": "INFORMATION_SCHEMA"
+                "warehouse": "COMPUTE_WH",
+                "database": "CORTEX_ANALYST_DEMO",
+                "schema": "REVENUE_TIMESERIES"
             }
         }
     ],
