+++ actions/snowflake-data/package.yaml
@@ -1,11 +1,11 @@
 # Required: A short name for the action package
-name: Snowflake Cortex
+name: Snowflake Data
 
 # Required: A description of what's in the action package.
-description: Snowflake Cortex AI Actions
+description: Various tools for interacting with Snowflake databases and other tools.
 
 # Package version number, recommend using semver.org
-version: 1.1.2
+version: 2.0.0
 
 spec-version: v2
 
@@ -14,17 +14,17 @@ dependencies:
   - python=3.10.16
   - uv=0.4.17
   pypi:
-  - sema4ai-actions=1.3.5
+  - sema4ai-actions=1.3.6
   - sema4ai-data=1.0.2
-  - snowflake-connector-python[pandas]=3.13.2
+  - snowflake-connector-python[pandas]=3.14.0
   - pandas=2.2.3
   - tabulate=0.9.0
   - PyYAML=6.0.2
   - openpyxl=3.2.0b1
-  - snowflake=1.0.5
-  - snowflake-snowpark-python=1.28.0
+  - snowflake=1.1.0
+  - snowflake-snowpark-python=1.29.1
   - pyjwt=2.10.1
-  - streamlit=1.42.2
+  - streamlit=1.43.2
 
 packaging:
   # By default, all files and folders in this directory are packaged when uploaded.
